STRAVA処理を記述しているコントローラを認証とそれ以外で分割
